requirements.txt
tensorflow-gpu==2.0.0
matplotlib==3.1.0
sklearn
seaborn
pandas